pyproject.toml
[project]
name = "tradingagents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=1.5",
    "yfinance>=0.2.37",
    "typer>=0.9",
    "rich>=13",
    "openai>=1.40.0",
    "google-generativeai>=0.7.0",
]
